Refactor: tests moved to circuitbreaker_helper.rs
Two tests remain with the block parsers, to handle the None case
and the failout
