+++ crates/config/src/kdl/routes.rs
@@ -1672,7 +1672,7 @@ fn parse_pii_detection_config(node: &kdl::KdlNode) -> Result<PiiDetectionConfig>
 #[cfg(test)]
 mod tests {
     use super::*;
-    use zentinel_common::{types::Priority, CircuitBreakerConfig};
+    use zentinel_common::types::Priority;
 
     /// Parse a KDL fragment like `route "test" { priority ... }` and return
     /// the resulting `Priority`. The parser expects a `route` parent node, so
@@ -1771,6 +1771,7 @@ mod tests {
     }
 
     /// circuit-breaker stanza present, all values normally set, use those values
+    /// Retain this test here to ensure block parser works
     #[test]
     fn test_parse_circuit_breaker_normal() {
         let kdl = r#"
@@ -1797,89 +1798,6 @@ mod tests {
         assert_eq!(cbconfig.success_threshold, 2);
         assert_eq!(cbconfig.timeout_seconds, 4);
         assert_eq!(cbconfig.half_open_max_requests, 8);
-    }
-
-    /// circuit-breaker stanza present, one key unrecognized, expect to Err and panic out
-    #[test]
-    fn test_parse_circuit_breaker_badkey() {
-        let kdl = r#"
-        routes {
-            route "test-cb" {
-                upstream "backend"
-                circuit-breaker {
-                    failure-threshold 1
-                    success-threshold 2
-                    timeout-sekonds 4
-                    half-open-max-requests 8
-                }
-            }
-        } 
-        "#;
-        let doc: kdl::KdlDocument = kdl.parse().unwrap();
-        let routes_node = doc.get("routes").unwrap();
-        let routes = parse_routes(routes_node);
-
-        let err_msg = routes.unwrap_err();
-
-        assert_eq!(format!("{}", err_msg), "Got unknown key timeout-sekonds");
-    }
-
-    /// circuit-breaker stanza present, new key unrecognized, expect to Err and panic out
-    #[test]
-    fn test_parse_circuit_breaker_badnewkey() {
-        let kdl = r#"
-        routes {
-            route "test-cb" {
-                upstream "backend"
-                circuit-breaker {
-                    failure-threshold 1
-                    success-threshold 2
-                    timeout-seconds 4
-                    half-open-max-requests 8
-                    reticulate 24
-                }
-            }
-        } 
-        "#;
-
-        let doc: kdl::KdlDocument = kdl.parse().unwrap();
-        let routes_node = doc.get("routes").unwrap();
-        let routes = parse_routes(routes_node);
-
-        let err_msg = routes.unwrap_err();
-
-        assert_eq!(format!("{}", err_msg), "Got unknown key reticulate");
-    }
-
-    /// circuit-breaker stanza present, all values missing, defaults should be used
-    #[test]
-    fn test_parse_circuit_breaker_fields_missing() {
-        let kdl = r#"
-        routes {
-            route "test-cb" {
-                upstream "backend"
-                circuit-breaker {
-                }
-            }
-        } 
-        "#;
-
-        let doc: kdl::KdlDocument = kdl.parse().unwrap();
-        let routes_node = doc.get("routes").unwrap();
-        let routes = parse_routes(routes_node).unwrap();
-        let route = routes.get(0).unwrap();
-
-        let cbconfig = route.circuit_breaker.unwrap();
-
-        let cb_default = CircuitBreakerConfig::default();
-
-        assert_eq!(cbconfig.failure_threshold, cb_default.failure_threshold);
-        assert_eq!(cbconfig.success_threshold, cb_default.success_threshold);
-        assert_eq!(cbconfig.timeout_seconds, cb_default.timeout_seconds);
-        assert_eq!(
-            cbconfig.half_open_max_requests,
-            cb_default.half_open_max_requests
-        );
     }
 
     /// circuit-breaker stanza missing, Option<CircuitBreakerConfig> will be None, upstream to set defaults
